requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Excel integration
openpyxl==3.1.2
xlsxwriter==3.1.9
pywin32==306; sys_platform == "win32"

# AI/NLP
openai==1.3.8
anthropic==0.7.8
tiktoken==0.5.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Web UI
jinja2==3.1.2
aiofiles==23.2.1

# Utilities
requests==2.31.0
rich==13.7.0
click==8.1.7
pyyaml==6.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1 